06/26/2016 new class LocationType added

+++ crashInTheCaribbean/src/CIT260/crashInTheCaribbean/model/Map.java
@@ -48,7 +48,7 @@ public class Map implements Serializable{
         return rowCount;
     }
 
-    public void setRowCount(double rowCount) {
+    public void setRowCount(int rowCount) {
         this.rowCount = rowCount;
     }
 
@@ -56,7 +56,7 @@ public class Map implements Serializable{
         return columnCount;
     }
 
-    public void setColumnCount(double columnCount) {
+    public void setColumnCount(int columnCount) {
         this.columnCount = columnCount;
     }
